Revert "Revert "[P5Lab] Move and rename animation metadata scripts""
