+++ bin/animation_assets/constants.rb
