+++ bin/animation_assets/for_each_library_animation.rb
