+++ bin/animation_assets/manifest_builder.rb
