+++ bin/animation_assets/rebuild_default_sprites.rb
